Refactor template for edit for DRY
